Reorganize plugins in a dedicated package
This commit moves all plugins into a `plugins` sub directory and
modifies required import statements. ST2 style imports are removed as
other changes (`embed` in syntax definitions, ...) of the package break
ST2 compatibility anyway.

The main reasons for this change are:

1. Some modules import others (e.g. nearly all import `mdeutils`). 
   As ST handles any top-level python module as independent plugin,
   importing one into another easily creates duplicate instances of
   those modules in python's runtime environment. As a result, reloading
   plugins (during update or development) may easily cause multiple
   instances of the same global object.

   This is no longer the case now.

2. The single `plugin.py` module in package root controls, which
   resources (API classes) are published to ST, so it doesn't need to
   query all modules one after another on its own.

   So we get more control about what ST sees or not.
   
   It hides the base class MDETextCommand for instance, which is of no
   value when called directly, but was registered as valid command
   before this commit.

3. All files are merged and ST just sees a single plugin module it needs
   to interact with, which makes loading/reloading faster.

The following things need to be kept in mind

1. Each command or event listener needs to be imported by plugin.py, 
   which means slightly higher effort when adding API functions.

2. `plugin.py` clears `sys.modules` cache upon import time to make sure
   all modules in the plugins directory are reloaded in correct order.

3. Changes to a module in the plugins sub directory take effect only by
   saving the main `plugin.py` (which causes everything to be reloaded).
