--- list_back_links.py
@@ -1,22 +1,0 @@
-import sublime, sublime_plugin
-import os, string
-
-try:
-    from MarkdownEditing.wiki_page import *
-except ImportError:
-    from wiki_page import *
-
-try:
-    from MarkdownEditing.mdeutils import *
-except ImportError:
-    from mdeutils import *
-
-
-class ListBackLinksCommand(MDETextCommand):
-    def run(self, edit):
-        print("Running ListBackLinksCommand")        
-        wiki_page = WikiPage(self.view)
-
-        file_list = wiki_page.find_files_with_ref()
-        wiki_page.select_backlink(file_list)
-
