+++ plugins/mdeutils.py
@@ -1,4 +1,3 @@
-import sublime
 import sublime_plugin
