+++ plugins/quote_indenting.py
@@ -1,10 +1,7 @@
 import re
 import sublime
-import sublime_plugin
-try:
-    from MarkdownEditing.mdeutils import *
-except ImportError:
-    from mdeutils import *
+
+from .mdeutils import *
 
 
 class IndentQuote(MDETextCommand):
